move check for linopy solver to grid_utils and adapt tests
