+++ assume/common/grid_utils.py
@@ -2,13 +2,17 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
+import logging
 from datetime import timedelta
 
 import numpy as np
 import pandas as pd
 import pypsa
+from linopy import available_solvers
 
 from assume.common.market_objects import MarketProduct
+
+logger = logging.getLogger(__name__)
 
 
 def add_generators(
@@ -333,3 +337,20 @@ def calculate_network_meta(network, product: MarketProduct, i: int):
         )
 
     return meta
+
+
+def get_supported_solver_linopy(default_solver: str | None = None):
+    SOLVERS = ["highs", "gurobi", "glpk", "cbc", "cplex"]
+
+    # Check which solver is available and filter to include only those in SOLVERS, preserving the order in SOLVERS
+    solvers = [solver for solver in SOLVERS if solver in available_solvers]
+    if not solvers:
+        raise RuntimeError(f"None of {SOLVERS} are available")
+
+    solver = default_solver or solvers[0]
+
+    if solver not in solvers:
+        logger.warning("Solver %s not available, using %s", solver, solvers[0])
+        solver = solvers[0]
+
+    return solver
