+++ tests/test_utils.py
@@ -22,7 +22,6 @@ from assume.common.utils import (
     datetime2timestamp,
     get_available_products,
     get_products_index,
-    get_supported_solver_linopy,
     get_supported_solver_pyomo,
     initializer,
     load_index_file,
@@ -815,12 +814,9 @@ def test_parse_duration():
         parse_duration("100ms")
 
 
-@pytest.mark.require_network
 def test_solver_available():
     assert get_supported_solver_pyomo() == "appsi_highs"
     assert get_supported_solver_pyomo("unknown_solver") == "appsi_highs"
-    assert get_supported_solver_linopy() == "highs"
-    assert get_supported_solver_linopy("unknown_solver") == "highs"
 
 
 def test_solver_unavailable(monkeypatch):
